Feature: Data limiting
Backend responses are now limited to a maximum of 1k records.

The BE gets the date range from the GET parameters, filters the data, works
out the minimum supported subtotal size that would fit within the limit,
then condenses the data accordingly.

Changing the date range sometimes fetches fresh data, since a smaller
date range may permit smaller subtotal groupings, offering more
granularity.

The "sizes in hours" for the subtotals are limited to a handful of
options to help with caching the data.

To support the stats cards, subtotals also store some useful details
that would otherwise be destroyed: start and end timestamps, details
about the min and max in each subtotal, and the number of individual
records it represents.

Also:
* Fixed issue with Plotly margins being too wide on mobile
* Removed use of useSWR's "isLoading" because it was causing a reload on
  every fetch
* To keep things DRY, the backend response now declares the number
  of records per hour
* Backend response now includes the min and max dates since the response
  data is no longer guaranteed to cover the entire available time range