Fix: Stats cards flickering / reset data broken
Bug #1:
On loading the page, the stats cards flickered and changed some of the
stats.

Cause:
Date and duration form.
The flickering was caused by it initialising its date to the minimum,
which counted as a change, which caused it to auto-update the date
range, which altered it because date and duration chops "straggler" data
off the ends of the range.

Factors to consider:
If the form initialises to "00:01", to capture a full day, it changes
the date range and so makes the stats flicker.

If the form invisibly initialises to "00:00" (to prevent flicker) then
it will give different results for the first day depending on if the
user uses it as it loads ("00:00"), or changes it to that date ("00:01")

Solution:
* Removed the automatic initialisation to the start date. It makes the
  date selection a bit more annoying, but avoids flicker and makes it
  behave consistently

----

Bug #2:
On resetting the data, the whole page crashed.

Cause:
Date and duration form.
The reset was also auto-triggering a date range change, which clashed
with the "main" reset, causing it to get its fetches in a twist and
fall on its face.

Solution:
Added a lock to prevent the auto-update when called by onReset()